+++ glide3x/sst1/glide3/src/gstrip.c
@@ -151,7 +151,7 @@
 #include "fxglide.h"
 
 /* access a floating point array with a byte index */
-#define FARRAY(p,i) (*(float *)((i)+(int)(p)))
+#define FARRAY(p,i) (*(float *)((i)+(unsigned long)(p)))
 /* access a byte array with a byte index and convert to float */
 #define FbARRAY(p,i) (float)(((unsigned char *)p)[i])
